was still doing pathfinding might have to restart it

+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: amalangu <amalangu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/21 15:19:23 by amalangu          #+#    #+#             */
-/*   Updated: 2025/01/21 23:15:30 by amalangu         ###   ########.fr       */
+/*   Updated: 2025/01/22 09:55:15 by amalangu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,10 +26,10 @@ void	print_array(t_map *map)
 	tmp = map->collectibles;
 	while (tmp)
 	{
-		ft_printf("%d %d\n", tmp->x, tmp->y);
+		ft_printf("%d %d\n", (int)tmp->x, (int)tmp->y);
 		tmp = tmp->next_collectible;
 		i++;
 	}
-	ft_printf("%d %d\n", map->collectibles->x, map->collectibles->y);
+	ft_printf("%d %d\n", (int)map->collectibles->x, (int)map->collectibles->y);
 	ft_printf("%d", i);
 }
